Moved MeasurementDetails to its own .py file.
